requirements.txt
# ===== OCR Engines =====
pytesseract>=0.3.10
paddleocr>=2.7.0
vietocr>=0.3.13
python-doctr>=0.5.0

# ===== ML/DL Libraries =====
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0

# ===== Image Processing =====
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pdf2image>=1.16.0

# ===== Text Processing =====
python-Levenshtein>=0.21.0
unidecode>=1.3.0
regex>=2023.0.0

# ===== Web Framework =====
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6